show available users on switchuser page
